refactor: add cli.service and state.service
