+++ __tests__/unit/categories/js/index.spec.ts
@@ -1,5 +1,5 @@
 import js from '#src/categories/js/index.ts';
-import { jsCategoryState } from '#src/states/categories.ts';
+import { jsCategoryState } from '#src/services/state.service.ts';
 
 describe('index', () => {
   describe('js', () => {
