+++ src/categories/js/build-workflow/build-workflow.config.ts
@@ -1,4 +1,4 @@
-import { jsCategoryState } from '#src/states/categories.ts';
+import { jsCategoryState } from '#src/services/state.service.ts';
 import { defaultConfig } from '#src/categories/js/build-workflow/config/default.config.ts';
 
 export const [getConfig] = jsCategoryState.useConfigState({
