+++ src/categories/js/jest/jest.config.ts
@@ -1,4 +1,4 @@
-import { jsCategoryState } from '#src/states/categories.ts';
+import { jsCategoryState } from '#src/services/state.service.ts';
 import { defaultConfig } from '#src/categories/js/jest/config/default.config.ts';
 import { nodeTsConfig } from '#src/categories/js/jest/config/node-ts.config.ts';
 import { reactTsConfig } from '#src/categories/js/jest/config/react-ts.config.ts';
