+++ src/categories/js/lint-staged/lint-staged.config.ts
@@ -1,4 +1,4 @@
-import { jsCategoryState } from '#src/states/categories.ts';
+import { jsCategoryState } from '#src/services/state.service.ts';
 import { defaultConfig } from '#src/categories/js/lint-staged/config/default.config.ts';
 import { nodeTsConfig } from '#src/categories/js/lint-staged/config/node-ts.config.ts';
 import { reactTsConfig } from '#src/categories/js/lint-staged/config/react-ts.config.ts';
