+++ src/categories/js/stylelint/stylelint.config.ts
@@ -1,4 +1,4 @@
-import { jsCategoryState } from '#src/states/categories.ts';
+import { jsCategoryState } from '#src/services/state.service.ts';
 import { defaultConfig } from '#src/categories/js/stylelint/config/default.config.ts';
 import { reactTsConfig } from '#src/categories/js/stylelint/config/react-ts.config.ts';
