--- src/states/categories.ts
@@ -1,3 +1,0 @@
-import { createCategoryState } from '#src/utils/state.ts';
-
-export const jsCategoryState = createCategoryState('js', ['node:ts', 'react:ts']);
